# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Estudiantes",
        "description": "Documentación de la API para gestionar estudiantes",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
        "/api/estudiante": {
            "get": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Obtener todos los estudiantes",
                "operationId": "41cf9865b42ec68717ed6165cc99d641",
                "responses": {
                    "200": {
                        "description": "Lista de estudiantes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Estudiante"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Crear un nuevo estudiante",
                "operationId": "17e5ea8eb84e82df3761dd9faac7fd09",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EstudianteInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estudiante creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Estudiante creada exitosamente"
                                        },
                                        "musica": {
                                            "$ref": "#/components/schemas/Estudiante"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/estudiante/{id}": {
            "get": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Obtener un estudiante por ID",
                "operationId": "b3f743b9edbb8f9aeadc9d192b20ce86",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estudiante encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Estudiante"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Estudiante no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Estudiante no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Actualizar un estudiante existente",
                "operationId": "2aa630b96bb7b647cf9b99b979110c7a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EstudianteInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estudiante actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Estudiante actualizado"
                                        },
                                        "Estudiante": {
                                            "$ref": "#/components/schemas/Estudiante"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Estudiante no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Estudiante no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Eliminar un estudiante",
                "operationId": "e8a355d1d176dd912d65d577ea541951",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estudiante eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Estudiante eliminada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Estudiante no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Estudiante no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Estudiante": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombres": {
                        "type": "string",
                        "example": "Juan"
                    },
                    "apellidos": {
                        "type": "string",
                        "example": "Pérez"
                    },
                    "cedula": {
                        "type": "string",
                        "example": "1234567890"
                    },
                    "correo": {
                        "type": "string",
                        "format": "email",
                        "example": "juan@email.com"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "555-1234"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "EstudianteInput": {
                "required": [
                    "nombres",
                    "apellidos",
                    "cedula",
                    "correo",
                    "telefono"
                ],
                "properties": {
                    "nombres": {
                        "type": "string",
                        "example": "Juan"
                    },
                    "apellidos": {
                        "type": "string",
                        "example": "Pérez"
                    },
                    "cedula": {
                        "type": "string",
                        "example": "1234567890"
                    },
                    "correo": {
                        "type": "string",
                        "format": "email",
                        "example": "juan@email.com"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "555-1234"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Estudiantes",
            "description": "Estudiantes"
        }
    ]
}